math: Intersect - fix typed comparison warnings

+++ spn/src/math/Intersect.cpp
@@ -2,12 +2,12 @@
 
 namespace math
 {
-	bool Intersect::pointRect( uint32_t pointX, uint32_t pointY,
-			uint32_t rectX, uint32_t rectY, uint32_t rectWidth, uint32_t rectHeight )
+	bool Intersect::pointRect( int32_t pointX, int32_t pointY,
+			int32_t rectX, int32_t rectY, uint32_t rectWidth, uint32_t rectHeight )
 	{
-		if ( pointX > ( rectX + rectWidth ) ||
+		if ( pointX > ( rectX + static_cast< int32_t >( rectWidth ) ) ||
 			 pointX < rectX ||
-			 pointY > ( rectY + rectHeight ) ||
+			 pointY > ( rectY + static_cast< int32_t >( rectHeight ) ) ||
 			 pointY < rectY )
 		{
 			return false;
